fix(state_machine_generator):fixed data init state
until now the init state used a string replace to put its code into
the execution state. since the execution state default content changed
this didn't work anymore. To avoid this in the future, the
code is now completely set into the execution state, and the default
content is removed (in stead of adjusted).